[connector] complete converter class
